refactor: restructure containment components and update imports
Moved containment components to a new directory structure under `components`. Updated all related imports to use the new paths and aliases, improving code organization and readability.
