+++ src/components/containment/carousel/carousel.component.tsx
