+++ src/components/containment/carousel/index.ts
