+++ src/components/containment/carousel/item/index.ts
