+++ src/components/containment/carousel/item/item.component.tsx
