--- src/containment/card/card.interface.ts
@@ -1,19 +1,0 @@
-import { ComponentProps } from '../../utils';
-import { ReactNode } from 'react';
-
-export type CardBaseProps = {
-
-  variant?: 'outlined' | 'elevated' | 'filled';
-  isInteractive?: boolean;
-};
-export type CardStates = {};
-export type CardElements =  'card' | 'stateLayer';
-export type CardElementType = 'div';
-
-export type CardProps = CardBaseProps &
-  ComponentProps<
-    CardBaseProps,
-    CardStates,
-    CardElements,
-    CardElementType
-  >;
