+++ src/index.ts
@@ -3,8 +3,8 @@ export * from './components/action/fab';
 export * from './components/action/icon-button';
 export * from './components/communication/progress-indicator';
 export * from './components/communication/snackbar';
-export * from './containment/card';
-export * from './containment/carousel';
+export * from '@components/containment/card';
+export * from '@components/containment/carousel';
 export * from './diviser';
 export * from './icon';
 export * from './navigation/tabs';
